# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hikāya API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema: { type: string }
        - name: state
          in: query
          required: false
          schema: { type: string }
        - name: iss
          in: query
          required: false
          schema: { type: string, format: uri }
      responses:
        '302':
          description: Redirect.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to OIDC logout.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Mobile logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /me/preferences:
    patch:
      tags: [me]
      operationId: updateMyPreferences
      summary: Update preferred UI language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesBody'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
...
  /me/progress:
    get:
      tags: [me]
      operationId: listMyProgress
      summary: Continue-reading list
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProgressEntry'
    post:
      tags: [me]
      operationId: upsertMyProgress
      summary: Save reading progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProgressBody'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressEntry'
  /me/bookmarks:
    get:
      tags: [me]
      operationId: listMyBookmarks
      summary: Bookmarked stories
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Story'
    post:
      tags: [me]
      operationId: toggleBookmark
      summary: Toggle bookmark on a story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToggleBookmarkBody'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToggleBookmarkResult'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, role, preferredLanguage]
      properties:
        id: { type: string }
        email: { type: [string, 'null'], format: email }
        firstName: { type: [string, 'null'] }
        lastName: { type: [string, 'null'] }
        profileImageUrl: { type: [string, 'null'] }
        role: { type: string }
        preferredLanguage: { type: string }
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code: { type: string, minLength: 1 }
        code_verifier: { type: string, minLength: 1 }
        redirect_uri: { type: string, minLength: 1, format: uri }
        state: { type: string, minLength: 1 }
        nonce: { type: string, minLength: 1 }
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token: { type: string }
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success: { type: boolean, const: true }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
    UpdatePreferencesBody:
      type: object
      required: [preferredLanguage]
      properties:
        preferredLanguage:
          type: string
    Story:
      type: object
      required: [id, slug, title, description, language, category, type, tags, status, accentColor, createdAt]
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        description: { type: string }
        coverImage: { type: [string, 'null'] }
        language: { type: string }
        category: { type: string }
        type: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { type: string }
        videoUrl: { type: [string, 'null'] }
        accentColor: { type: string }
        createdAt: { type: string, format: date-time }
        chapterCount: { type: integer }
...
    ProgressEntry:
      type: object
      required: [storyId, chapterNumber, progressPercent, updatedAt, story]
      properties:
        storyId: { type: string }
        chapterNumber: { type: integer }
        progressPercent: { type: integer }
        updatedAt: { type: string, format: date-time }
        story: { $ref: '#/components/schemas/Story' }
    UpsertProgressBody:
      type: object
      required: [storyId, chapterNumber, progressPercent]
      properties:
        storyId: { type: string }
        chapterNumber: { type: integer }
        progressPercent: { type: integer }
    ToggleBookmarkBody:
      type: object
      required: [storyId]
      properties:
        storyId: { type: string }
    ToggleBookmarkResult:
      type: object
      required: [bookmarked]
      properties:
        bookmarked: { type: boolean }
...