# lib/api-zod/src/generated/types/progressEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hikāya API
 * OpenAPI spec version: 0.1.0
 */
import type { Story } from "./story";

export interface ProgressEntry {
  storyId: string;
  chapterNumber: number;
  progressPercent: number;
  updatedAt: Date;
  story: Story;
}